requirements_local.txt
Flask==2.2.5
Werkzeug==2.2.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8

Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.44
PyMySQL==1.1.2

Flask-Migrate==4.1.0
alembic==1.16.5
Mako==1.3.10

python-dotenv==1.2.1

Flask-Caching==2.3.1
cachelib==0.13.0

PyLTI1p3==2.0.0

google-generativeai==0.8.6
google-api-python-client==2.188.0
google-auth==2.47.0
 